"#5: Creada la función de añadir puntos"
Se ha creado una api con los negocios y funciones para convertir el
texto de la dirección el coordenadas.
Para mostrarlos en el mapa se ha usado una función javascript que parsea
el json traido desde la api a un geojson

+++ src/main/java/com/gbweb/entity/Localizacion.java
@@ -30,13 +30,13 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 public class Localizacion {
 
     @JsonProperty("place_id")
-    private Integer placeId;
+    private Long placeId;
     @JsonProperty("licence")
     private String licence;
     @JsonProperty("osm_type")
     private String osmType;
     @JsonProperty("osm_id")
-    private Integer osmId;
+    private Long osmId;
     @JsonProperty("boundingbox")
     private List<String> boundingbox = null;
     @JsonProperty("lat")
@@ -55,12 +55,12 @@ public class Localizacion {
     private Map<String, Object> additionalProperties = new HashMap<String, Object>();
 
     @JsonProperty("place_id")
-    public Integer getPlaceId() {
+    public Long getPlaceId() {
         return placeId;
     }
 
     @JsonProperty("place_id")
-    public void setPlaceId(Integer placeId) {
+    public void setPlaceId(Long placeId) {
         this.placeId = placeId;
     }
 
@@ -85,12 +85,12 @@ public class Localizacion {
     }
 
     @JsonProperty("osm_id")
-    public Integer getOsmId() {
+    public Long getOsmId() {
         return osmId;
     }
 
     @JsonProperty("osm_id")
-    public void setOsmId(Integer osmId) {
+    public void setOsmId(Long osmId) {
         this.osmId = osmId;
     }
